pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.34.2",
    "ffmpeg>=1.4",
    "ipython>=8.27.0",
    "nltk>=3.9.1",
    "numpy>=2.1.1",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scipy>=1.14.1",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.1",
    "suno-bark",
    "tk>=0.1.0",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "transformers>=4.44.2",
]

[tool.uv.sources]
suno-bark = { git = "https://github.com/suno-ai/bark.git" }
